Add test_base_gates
