+++ src/qibo/tests/test_decompositions.py
@@ -1,49 +1,10 @@
-import re
 import numpy as np
 import pytest
-import cirq
 from qibo import gates
 from qibo.models import Circuit
 from qibo.tests import utils
 
-_QIBO_TO_CIRQ = {"CNOT": "CNOT", "RY": "Ry", "TOFFOLI": "TOFFOLI"}
 _ATOL = 1e-6
-
-
-def assert_gates_equivalent(qibo_gate, cirq_gate):
-    """Asserts that qibo gate is equivalent to cirq gate.
-
-    Checks that:
-        * Gate type agrees.
-        * Target and control qubits agree.
-        * Parameter (if applicable) agrees.
-
-    Cirq gate parameters are extracted by parsing the gate string.
-    """
-    pieces = [x for x in re.split("[()]", str(cirq_gate)) if x]
-    if len(pieces) == 2:
-        gatename, targets = pieces
-        theta = None
-    elif len(pieces) == 3:
-        gatename, theta, targets = pieces
-    else: # pragma: no cover
-        # case not tested because it fails
-        raise RuntimeError("Cirq gate parsing failed with {}.".format(pieces))
-
-    qubits = list(int(x) for x in targets.replace(" ", "").split(","))
-    targets = (qubits.pop(),)
-    controls = set(qubits)
-
-    assert _QIBO_TO_CIRQ[qibo_gate.__class__.__name__] == gatename
-    assert qibo_gate.target_qubits == targets
-    assert set(qibo_gate.control_qubits) == controls
-    if theta is not None:
-        if "π" in theta:
-            theta = np.pi * float(theta.replace("π", ""))
-        else: # pragma: no cover
-            # case doesn't happen in tests (could remove)
-            theta = float(theta)
-        np.testing.assert_allclose(theta, qibo_gate.parameters)
 
 
 def assert_circuit_same_gates(circuit1, circuit2):
@@ -58,27 +19,6 @@ def assert_circuit_same_gates(circuit1, circuit2):
     if mg1 is not None:
         assert mg1.target_qubits == mg2.target_qubits
     assert circuit1.measurement_tuples == circuit2.measurement_tuples
-
-
-@pytest.mark.parametrize(("target", "controls", "free"),
-                         [(0, (1,), ()), (2, (0, 1), ()),
-                          (3, (0, 1, 4), (2, 5)),
-                          (7, (0, 1, 2, 3, 4), (5, 6))])
-def test_x_decomposition_gates(target, controls, free):
-    """Check that decomposition of multi-control ``X`` agrees with Cirq."""
-    gate = gates.X(target).controlled_by(*controls)
-    qibo_decomp = gate.decompose(*free, use_toffolis=False)
-
-    # Calculate the decomposition using Cirq.
-    nqubits = max((target,) + controls + free) + 1
-    qubits = [cirq.LineQubit(i) for i in range(nqubits)]
-    controls = [qubits[i] for i in controls]
-    free = [qubits[i] for i in free]
-    cirq_decomp = cirq.decompose_multi_controlled_x(controls, qubits[target], free)
-
-    assert len(qibo_decomp) == len(cirq_decomp)
-    for qibo_gate, cirq_gate in zip(qibo_decomp, cirq_decomp):
-        assert_gates_equivalent(qibo_gate, cirq_gate)
 
 
 @pytest.mark.parametrize(("target", "controls", "free"),
@@ -103,26 +43,6 @@ def test_x_decomposition_execution(target, controls, free, use_toffolis):
     final_state = c(np.copy(init_state)).numpy()
 
     np.testing.assert_allclose(final_state, target_state, atol=_ATOL)
-
-
-@pytest.mark.parametrize("use_toffolis", [True, False])
-def test_x_decomposition_errors(use_toffolis):
-    """Check ``X`` decomposition errors."""
-    gate = gates.X(0).controlled_by(1, 2, 3, 4)
-    with pytest.raises(ValueError):
-        decomp = gate.decompose(2, 3, use_toffolis=use_toffolis)
-    c = Circuit(6)
-    c.add(gate)
-    with pytest.raises(ValueError):
-        decomp = gate.decompose(5, 6, use_toffolis=use_toffolis)
-
-
-def test_x_decompose_few_controls():
-    """Check ``X`` decomposition with len(controls) < 3."""
-    gate = gates.X(0)
-    decomp = gate.decompose(1, 2)
-    assert len(decomp) == 1
-    assert isinstance(decomp[0], gates.X)
 
 
 def test_circuit_decompose():
